server/requirements.txt
Flask
psycopg2
requests
dnspython
PyJWT==2.9.0
python-dotenv
Flask-JWT-Extended